Refactoring (#2)
Update structure
Add Concurrency tests for LockID/UnlockID
