+++ Makefile
@@ -1,2 +1,5 @@
-test:
-	go test -cover
+tests:
+	go test -cover ./test -v
+
+samples:
+	go run ./cmd/sample/main.go
