+++ consts.go
@@ -1,0 +1,9 @@
+package simplejsondb
+
+import "errors"
+
+var (
+	Ext            string = ".json"
+	GZipExt        string = ".json.gz"
+	ErrNoDirectory error  = errors.New("not a directory")
+)
