+++ test/helpers_test.go
@@ -1,0 +1,22 @@
+package test_test
+
+import (
+	"os"
+	"path/filepath"
+
+	"golang.org/x/exp/rand"
+)
+
+const letterBytes = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ"
+
+func randName(n int) string {
+	b := make([]byte, n)
+	for i := range b {
+		b[i] = letterBytes[rand.Intn(len(letterBytes))]
+	}
+	return string(b)
+}
+
+func remove(dir ...string) error {
+	return os.Remove(filepath.Join(dir...))
+}
